clean up some of the self join stuff
